work on the live activity ui
