+++ app/src/app/(tabs)/(profile)/profile/cars/index.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { ActivityIndicator, FlatList } from "react-native";
 import { PAGE_SIZE } from "@/utils/constants";
 import { useUser } from "@/utils/useUser";
@@ -12,24 +11,8 @@ import { keepPreviousData, useInfiniteQuery } from "@tanstack/react-query";
 import { useMutation } from "@tanstack/react-query";
 import { useQueryClient } from "@tanstack/react-query";
 import { Menu } from "@/components/Menu";
-import { Stack, useRouter } from "expo-router";
 import { getContentContainerStyle } from "@/utils/styles";
-
-const colorMap = {
-  red: "#ca3f3f",
-  green: "#62be62",
-  blue: "#4285ea",
-  purple: "#a837b7",
-  black: "#2b2b2b",
-  gray: "#a8a8a8",
-  pink: "#d36ecb",
-  white: "#e2e2e2",
-  orange: "#d8670a",
-  tan: "#c69567",
-  brown: "#78513edd",
-  silver: "#7e7e7e",
-  yellow: "#ffc72f",
-};
+import { colorMap } from "@/utils/cars";
 
 export default function Cars() {
   const trpc = useTRPC();
